Replaced - (unsigned int) count by - (NSUInteger) count in all frameworks to prevent return type conflict with GNUstep Foundation classes.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@6128 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/CollectionKit/Headers/CKMultiValue.h
@@ -34,7 +34,7 @@
 - (id) initWithType: (CKPropertyType) type contentArray: (NSArray *) array;
 - (NSArray*) contentArray;
 
-- (unsigned int) count;
+- (NSUInteger) count;
 
 - (id) valueAtIndex: (int) index;
 - (NSString*) labelAtIndex: (int) index;
