+++ Frameworks/CollectionKit/Source/CKMultiValue.m
@@ -93,7 +93,7 @@ static CKPropertyType _propTypeFromDict(NSDictionary *dict)
   return _arr;
 }
 
-- (unsigned int) count
+- (NSUInteger) count
 {
   return [_arr count];
 }
